[plugins] Don't move RouteOverviewUtil in this PR
